Validações e Testes feitos e completos no contexto de cenarios

+++ docker-django/src/cenarios/views.py
@@ -2,9 +2,10 @@ from django.shortcuts import render,redirect,get_object_or_404
 from .models import Insumo,Cenario,Produto 
 from .forms import InsumoForm,ProdutoForm,CenarioForm
 from django.http import HttpRequest
-# Create your views here.
+from django.contrib import messages
+from jogos.models import Jogo
 
-from django.http import HttpResponse
+# Create your views here.
 
 #lógica dos CRUDs em uma unica função da view
 def cenarios_view(request:HttpRequest):
@@ -12,17 +13,6 @@ def cenarios_view(request:HttpRequest):
     insumo_form = InsumoForm()
     produto_form = ProdutoForm()
     cenario_form = CenarioForm()
-
-
-    #lista todos cenarios,produtos e insumo
-    contexto = { 
-        "insumos": Insumo.objects.all(),
-        "cenarios": Cenario.objects.all(),
-        "produtos": Produto.objects.all(),
-        "insumo_form": insumo_form,
-        "produto_form": produto_form,
-        "cenario_form": cenario_form,
-    }
 
     if request.method == "POST":
         model_type = request.POST.get("model_type")
@@ -33,41 +23,77 @@ def cenarios_view(request:HttpRequest):
                 formInsumo = InsumoForm(request.POST)
                 if formInsumo.is_valid():
                     formInsumo.save()
+                    messages.success(request, "Insumo criado com sucesso!")
                     return redirect("cenarios:home")
+                else:
+                    if formInsumo.errors.get('nome'):
+                        messages.error(request, "Erro ao salvar insumo, nome do insumo inválido!")
+                    if formInsumo.errors.get('fornecedor'):
+                        messages.error(request,"Erro ao salvar insumo, nome do fornecedor inválido")
+                    insumo_form = formInsumo
                 
         if model_type == "produto":
             if action == "create":
                 formProduto = ProdutoForm(request.POST)
                 if formProduto.is_valid():
                     formProduto.save()
+                    messages.success(request, "Produto criado com sucesso!")
                     return redirect("cenarios:home")
+                else:
+                    if formProduto.errors.get('nome'):
+                        messages.error(request,"Erro ao salvar Produto, nome do Produto Inválido")
+                    if formProduto.errors.get('insumos'):
+                        messages.error(request,"Erro ao salvar, Produto sem insumos!")
+                    produto_form = formProduto
                 
         if model_type == "cenario":
             if action == "create":
                 formCenario = CenarioForm(request.POST)
                 if formCenario.is_valid():
                     formCenario.save()
+                    messages.success(request,"Cenário criado com sucesso!")
                     return redirect("cenarios:home")
+                else:
+                    if formCenario.errors.get('nome'):
+                        messages.error(request,"Erro ao salvar Cenário, nome do Cenário Inválido")
+                    if formCenario.errors.get('produto'):
+                        messages.error(request,"Erro ao salvar, é necessário um produto vinculado para criar um Cenário!")
                 
+    #lista todos cenarios,produtos e insumo e seus respectivos forms
+    contexto = { 
+        "insumos": Insumo.objects.all(),
+        "cenarios": Cenario.objects.all(),
+        "produtos": Produto.objects.all(),
+        "insumo_form": insumo_form,
+        "produto_form": produto_form,
+        "cenario_form": cenario_form,
+    }
+                      
     return render(request,'cenarios/cenarios.html',contexto)
-
-
-
 
 def removerInsumo(request:HttpRequest,id):
     insumo = get_object_or_404(Insumo,id=id)
     insumo.delete()
+    messages.success(request,"Insumo Deletado com sucesso")
     return redirect("cenarios:home")
 
 def removerProduto(request:HttpRequest,id):
     produto = get_object_or_404(Produto,id=id)
     produto.delete()
+    messages.success(request,"Produto Deletado com sucesso")
     return redirect("cenarios:home")
 
 def removerCenario(request:HttpRequest,id):
-    #logica de ver se o cenario está em um jogo ativo
+
     cenario = get_object_or_404(Cenario,id=id)
-    cenario.delete()
+    #logica de ver se o cenario está em um jogo ativo
+    jogoAtivo = Jogo.objects.filter(cenario=cenario, status = Jogo.ATIVO).exists()
+
+    if jogoAtivo:
+        messages.error(request,"Este Cenário está em um Jogo Ativo, Não é possível deletá-lo")
+    else:
+        cenario.delete()
+        messages.success(request,"Cenário Deletado com sucesso")
     return redirect("cenarios:home")
 
 def editarInsumo(request:HttpRequest, id):
@@ -76,7 +102,10 @@ def editarInsumo(request:HttpRequest, id):
         formInsumo = InsumoForm(request.POST,instance=insumo)
         if formInsumo.is_valid():
             formInsumo.save()
+            messages.success(request,"Insumo Editado com sucesso")
             return redirect("cenarios:home")
+        else:
+            messages.success(request,"Erro ao atualizar insumo!")
     formInsumo = InsumoForm(instance=insumo)
     contexto = { 
         'insumo_form':formInsumo,
@@ -93,7 +122,10 @@ def editarProduto(request:HttpRequest, id):
         formProduto = ProdutoForm(request.POST,instance=produto)
         if formProduto.is_valid():
             formProduto.save()
+            messages.success(request,"Produto Editado com sucesso")
             return redirect("cenarios:home")
+        else:
+            messages.error(request,"Erro ao atualizar Produto!")
     formProduto = ProdutoForm(instance=produto)
     contexto = {
         'produto_form' : formProduto,
@@ -110,7 +142,10 @@ def editarCenario(request:HttpRequest, id):
         formCenario = CenarioForm(request.POST, instance = cenario)
         if formCenario.is_valid():
             formCenario.save()
+            messages.success(request,"Cenário Editado com sucesso")
             return redirect ("cenarios:home")
+        else:
+            messages.error(request,"Erro ao editar Cenário!")
     formCenario = CenarioForm(instance=cenario)
     contexto = {
         'cenario_form' : formCenario,
